feat: 좋아요 추가/삭제, 좋아요 개수, 좋아요 여부

+++ controller/LikeController.js
@@ -1,34 +1,74 @@
 const conn = require('../mariadb');
 const { StatusCodes } = require('http-status-codes');
+const jwt = require('jsonwebtoken');
+const dotenv = require('dotenv');
+dotenv.config();
 
 const addLike = (req, res) => {
-  const { id } = req.params;
-  const { user_id } = req.body;
+  const book_id = req.params.id;
+  let authorization = ensureAuthorization(req, res);
 
-  let sql = `INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)`;
-  let values = [user_id, id];
-  conn.query(sql, values, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
-    return res.status(StatusCodes.OK).json(results);
-  });
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: `로그인 세션이 만료되었습니다. 다시 로그인 해주세요.`,
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: `잘못된 토큰입니다.`,
+    });
+  } else {
+    let sql = `INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)`;
+    let values = [authorization.id, book_id];
+    conn.query(sql, values, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
+      return res.status(StatusCodes.OK).json(results);
+    });
+  }
 };
 
 const cancelLike = (req, res) => {
-  const { id } = req.params;
-  const { user_id } = req.body;
+  const book_id = req.params.id;
+  let authorization = ensureAuthorization(req, res);
 
-  let sql = `DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?`;
-  let values = [user_id, id];
-  conn.query(sql, values, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
-    return res.status(StatusCodes.OK).json(results);
-  });
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: `로그인 세션이 만료되었습니다. 다시 로그인 해주세요.`,
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: `잘못된 토큰입니다.`,
+    });
+  } else {
+    let sql = `DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?`;
+    let values = [authorization.id, book_id];
+    conn.query(sql, values, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
+      return res.status(StatusCodes.OK).json(results);
+    });
+  }
 };
+
+function ensureAuthorization(req, res) {
+  try {
+    let receivedJwt = req.headers['authorization'];
+    console.log('receivedJwt : ', receivedJwt);
+
+    let decodedUserInfo = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
+    console.log(decodedUserInfo);
+
+    return decodedUserInfo;
+  } catch (err) {
+    console.log(err.name);
+    console.log(err.message);
+
+    return err;
+  }
+}
 
 module.exports = { addLike, cancelLike };
